Move `ConfigureBoxes` command
