+++ wcfsetup/install/files/lib/system/acp/dashboard/command/ConfigureBoxes.class.php
@@ -13,6 +13,7 @@ use wcf\system\WCF;
  * @copyright   2001-2023 WoltLab GmbH
  * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since       6.1
+ * @deprecated  6.2 Use `\wcf\command\acp\dashboard\ConfigureBoxes` instead.
  */
 final class ConfigureBoxes
 {
@@ -23,47 +24,10 @@ final class ConfigureBoxes
         private readonly AcpDashboard $dashboard,
         private readonly User $user,
         private readonly array $boxes,
-    ) {
-    }
+    ) {}
 
     public function __invoke(): void
     {
-        $this->resetBoxes();
-        $this->saveBoxes();
-    }
-
-    private function resetBoxes(): void
-    {
-        $sql = "DELETE FROM wcf1_acp_dashboard_box_to_user WHERE userID = ?";
-        $statement = WCF::getDB()->prepare($sql);
-        $statement->execute([$this->user->userID]);
-    }
-
-    private function saveBoxes(): void
-    {
-        $sql = "INSERT INTO wcf1_acp_dashboard_box_to_user (boxName, userID, enabled, showOrder) VALUES (?, ?, ?, ?)";
-        $statement = WCF::getDB()->prepare($sql);
-        $showOrder = 0;
-        foreach ($this->boxes as $boxName) {
-            $statement->execute([
-                $boxName,
-                $this->user->userID,
-                1,
-                $showOrder++
-            ]);
-        }
-
-        foreach ($this->dashboard->getBoxes() as $box) {
-            if (\in_array($box->getName(), $this->boxes)) {
-                continue;
-            }
-
-            $statement->execute([
-                $box->getName(),
-                $this->user->userID,
-                0,
-                0
-            ]);
-        }
+        (new \wcf\command\acp\dashboard\ConfigureBoxes($this->dashboard, $this->user, $this->boxes))();
     }
 }
